fix idempotent issue with systemd reload
there is no need for systemd to reload itself
unless the unit files on disk change.  since the
firewall class does not manipulate the unit file,
this exec statement is not needed, and in fact,
causes idempotent issues with ip6tables disabled.

see systemctl(1) for the details on daemon-reload
and what it does, non of which is required here.